sis-api/requirements.txt
connexion
flask
flask-injector
connexion[swagger-ui]
flask_testing==0.6.1
coverage>=4.0.3
nose>=1.3.7
pluggy>=0.3.1
py>=1.4.31
randomize>=0.13
behave
python-dateutil
pyhamcrest
encoder
passlib
pyyaml